+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingMerger.java
@@ -48,6 +48,7 @@ public class IndexingMerger {
     private final Index index;
     private long mergesLimit;
     private int mergeSuccesses = 0;
+    private long timeQuotaMillis = 0;
     private final IndexingCommon common;
 
     public IndexingMerger(final Index index,  IndexingCommon common, long initialMergesCountLimit) {
@@ -73,6 +74,7 @@ public class IndexingMerger {
                                     final IndexDeferredMaintenanceControl mergeControl = store.getIndexDeferredMaintenanceControl();
                                     mergeControlRef.set(mergeControl);
                                     mergeControl.setMergesLimit(mergesLimit);
+                                    mergeControl.setTimeQuotaMillis(timeQuotaMillis);
                                     return store.getIndexMaintainer(index).mergeIndex();
                                 }).thenApply(ignore -> false),
                         Pair::of,
@@ -80,54 +82,94 @@ public class IndexingMerger {
                 ).handle((ignore, e) -> {
                     recordTime.get().run();
                     final IndexDeferredMaintenanceControl mergeControl = mergeControlRef.get();
+                    // Note: this mergeControl will not be re-used and should not be modified.
                     if (e == null) {
-                        if (mergesLimit > 0 && mergeSuccesses > 2) {
-                            mergeSuccesses = 0;
-                            mergesLimit = (mergesLimit * 5) / 4; // increase 25%, case there was an isolated issue
-                        }
-                        mergeSuccesses++;
-                        if (LOGGER.isDebugEnabled()) {
-                            LOGGER.debug(KeyValueLogMessage.build("IndexMerge: Success")
-                                    .addKeysAndValues(mergerKeysAndValues(mergeControl))
-                                    .toString());
-                        }
-                        // Here: no error, stop the iteration unless has more
-                        final boolean hasMore = mergeControl.getMergesFound() > mergeControl.getMergesTried();
-                        return hasMore ? AsyncUtil.READY_TRUE : AsyncUtil.READY_FALSE;
+                        // Here: no errors
+                        return handleSuccess(mergeControl);
                     }
-                    // Here: got exception.
-                    if (0 > failureCountLimit.decrementAndGet() || mergeControl.getMergesTried() < 2) {
-                        if (LOGGER.isWarnEnabled()) {
-                            LOGGER.warn(KeyValueLogMessage.build("IndexMerge: Gave up merge dilution")
-                                            .addKeysAndValues(mergerKeysAndValues(mergeControl))
-                                            .toString(), e);
-                        }
-                    } else {
-                        final FDBException ex = IndexingBase.findException(e, FDBException.class);
-                        if (IndexingBase.shouldLessenWork(ex)) {
-                            // Here: this exception might be resolved by reducing the load
-                            mergesLimit = mergeControl.getMergesTried() / 2;
-                            if (LOGGER.isInfoEnabled()) {
-                                // TODO: demote this info message to a trace or debug after this code is tested a bit
-                                LOGGER.info(KeyValueLogMessage.build("IndexMerge: Merges diluted")
-                                        .addKeysAndValues(mergerKeysAndValues(mergeControl))
-                                        .toString(), e);
-                            }
-                            return AsyncUtil.READY_TRUE; // and retry
-                        }
+                    if (0 > failureCountLimit.decrementAndGet()) {
+                        // Here: too many retries, unconditionally give up
+                        giveUpMerging(mergeControl, e);
                     }
-                    // Here: this exception will not be recovered by dilution. Throw it.
-                    throw common.getRunner().getDatabase().mapAsyncToSyncException(e);
+                    return handleFailure(mergeControl, e);
                 }).thenCompose(Function.identity()
                 ), common.getRunner().getExecutor());
     }
 
+    private CompletableFuture<Boolean> handleSuccess(final IndexDeferredMaintenanceControl mergeControl) {
+        if (mergesLimit > 0 && mergeSuccesses > 2) {
+            mergeSuccesses = 0;
+            mergesLimit = (mergesLimit * 5) / 4; // increase 25%, case there was an isolated issue
+        }
+        mergeSuccesses++;
+        if (LOGGER.isDebugEnabled()) {
+            LOGGER.debug(KeyValueLogMessage.build("IndexMerge: Success")
+                    .addKeysAndValues(mergerKeysAndValues(mergeControl))
+                    .toString());
+        }
+        // after a successful merge, reset the time quota. It is unlikely to be applicable for the next merge.
+        timeQuotaMillis = 0;
+        // Here: no errors, stop the iteration unless has more
+        final boolean hasMore = mergeControl.getMergesFound() > mergeControl.getMergesTried();
+        return  hasMore ? AsyncUtil.READY_TRUE : AsyncUtil.READY_FALSE;
+    }
+
+    private CompletableFuture<Boolean> handleFailure(final IndexDeferredMaintenanceControl mergeControl, Throwable e) {
+        final FDBException ex = IndexingBase.findException(e, FDBException.class);
+        if (!IndexingBase.shouldLessenWork(ex)) {
+            giveUpMerging(mergeControl, e);
+        }
+        // Here: this exception might be resolved by reducing the number of merges or forcing shorter intervals between auto-commits
+        if (mergeControl.getMergesTried() < 2) {
+            handleSingleMergeFailure(mergeControl, e);
+        } else {
+            handleMultiMergeFailure(mergeControl, e);
+        }
+        return AsyncUtil.READY_TRUE; // and retry
+    }
+
+    private void handleMultiMergeFailure(final IndexDeferredMaintenanceControl mergeControl, Throwable e) {
+        // Here: reduce the number of OneMerge items attempted
+        mergesLimit = mergeControl.getMergesTried() / 2;
+        if (LOGGER.isTraceEnabled()) {
+            LOGGER.trace(KeyValueLogMessage.build("IndexMerge: Merges diluted")
+                    .addKeysAndValues(mergerKeysAndValues(mergeControl))
+                    .toString(), e);
+        }
+    }
+
+    private void handleSingleMergeFailure(final IndexDeferredMaintenanceControl mergeControl, Throwable e) {
+        // Here: make agility context auto-commit more rapidly
+        // Note: this will only change the time quota. Size quota seems to be a non-issue.
+        timeQuotaMillis = mergeControl.getTimeQuotaMillis();
+        // log 4000 base 2 =~ 11.96 So it'll take about 12 retries from 4 seconds to the minimum.
+        if (timeQuotaMillis <= 2) {
+            giveUpMerging(mergeControl, e);
+        }
+        timeQuotaMillis /= 2;
+        if (LOGGER.isTraceEnabled()) {
+            LOGGER.trace(KeyValueLogMessage.build("IndexMerge: Decrease time quota")
+                    .addKeysAndValues(mergerKeysAndValues(mergeControl))
+                    .toString(), e);
+        }
+    }
+
+    private void giveUpMerging(final IndexDeferredMaintenanceControl mergeControl, Throwable e) {
+        if (LOGGER.isWarnEnabled()) {
+            LOGGER.warn(KeyValueLogMessage.build("IndexMerge: Gave up merge dilution")
+                    .addKeysAndValues(mergerKeysAndValues(mergeControl))
+                    .toString(), e);
+        }
+        throw common.getRunner().getDatabase().mapAsyncToSyncException(e);
+    }
+
     List<Object> mergerKeysAndValues(final IndexDeferredMaintenanceControl mergeControl) {
         return List.of(
-                    LogMessageKeys.INDEX_NAME, index.getName(),
-                    LogMessageKeys.INDEX_MERGES_LIMIT, mergeControl.getMergesLimit(),
-                    LogMessageKeys.INDEX_MERGES_FOUND, mergeControl.getMergesFound(),
-                    LogMessageKeys.INDEX_MERGES_TRIED, mergeControl.getMergesTried()
+                LogMessageKeys.INDEX_NAME, index.getName(),
+                LogMessageKeys.INDEX_MERGES_LIMIT, mergeControl.getMergesLimit(),
+                LogMessageKeys.INDEX_MERGES_FOUND, mergeControl.getMergesFound(),
+                LogMessageKeys.INDEX_MERGES_TRIED, mergeControl.getMergesTried(),
+                LogMessageKeys.INDEX_MERGES_CONTEXT_TIME_QUOTA, mergeControl.getTimeQuotaMillis()
         );
     }