+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingMutuallyByRecords.java
@@ -441,7 +441,7 @@ public class IndexingMutuallyByRecords extends IndexingBase {
                                           lastResult.get().get().getPrimaryKey() :
                                           rangeEnd)
                     .thenCompose(cont -> insertRanges(targetRangeSets, packOrNull(rangeStart), packOrNull(cont))
-                            .thenApply(ignore -> !allRangesExhausted(cont, rangeEnd)));
+                            .thenApply(ignore -> notAllRangesExhausted(cont, rangeEnd)));
         });
     }