+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/provider/foundationdb/IndexingScrubDangling.java
@@ -178,7 +178,7 @@ public class IndexingScrubDangling extends IndexingBase {
                                         return false;
                                     }
                                 }
-                                return !allRangesExhausted(cont, rangeEnd);
+                                return notAllRangesExhausted(cont, rangeEnd);
                             }));
         });
     }