+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/values/OfTypeValue.java
@@ -26,6 +26,7 @@ import com.apple.foundationdb.record.ObjectPlanHash;
 import com.apple.foundationdb.record.PlanHashable;
 import com.apple.foundationdb.record.provider.foundationdb.FDBRecordStoreBase;
 import com.apple.foundationdb.record.query.plan.cascades.AliasMap;
+import com.apple.foundationdb.record.query.plan.cascades.PromoteValue;
 import com.apple.foundationdb.record.query.plan.cascades.typing.Type;
 import com.google.protobuf.DynamicMessage;
 import com.google.protobuf.Message;
@@ -86,7 +87,12 @@ public class OfTypeValue extends AbstractValue implements Value.RangeMatchableVa
             return expectedType.isRecord();
         }
         final var type = Type.fromObject(value);
-        return expectedType.equals(type);
+        final var promotionNeeded = PromoteValue.isPromotionNeeded(type, expectedType);
+        if (!promotionNeeded) {
+            return true;
+        }
+        final var promotionFunction = PromoteValue.resolvePromotionFunction(type, expectedType);
+        return promotionFunction != null;
     }
 
     @Override
@@ -116,6 +122,12 @@ public class OfTypeValue extends AbstractValue implements Value.RangeMatchableVa
         return new OfTypeValue(value, type);
     }
 
+    /**
+     * Derives a {@link OfTypeValue} object from a given {@link ConstantObjectValue}. It does this be constructing a new
+     * {@link ConstantObjectValue} as a child requiring it to have a type that conforms to the type of the passed {@link ConstantObjectValue}.
+     * @param value The {@link ConstantObjectValue} object we want to derive from.
+     * @return New {@link OfTypeValue} that checks whether the underlying child have a type conforming to the type of the {@link ConstantObjectValue}.
+     */
     @Nonnull
     public static OfTypeValue from(@Nonnull final ConstantObjectValue value) {
         return new OfTypeValue(ConstantObjectValue.of(value.getAlias(), value.getOrdinal(), Type.any()), value.getResultType());